refactor(cli): extract watch dispatch from cli/mod.rs into cli/watch.rs
The `Command::Watch` arm in `cli/mod.rs::execute` had ballooned to 173
lines of inline parsing + polling + condition-checking — the largest single
arm in the dispatch table, and the only one that hadn't already been moved
to its own module. Lift it into `src/cli/watch.rs` and reduce the arm to
a single call site.

Structurally the extracted code is split into four small functions instead
of one big block:

  * `execute_watch` — top-level dispatch on `target` vs `since`
  * `watch_until`   — `--target` mode: parse `key=<op><value>`, poll until met
  * `watch_since`   — `--since` mode: diff cached state-at vs state-now
  * `check_guest_condition` / `check_storage_condition` — per-target probes

This is the only change beyond the literal lift. Three minor cleanups fall
out naturally from splitting:

  * `if let Ok(vmid) = vmid_str.parse::<u32>() { ... } else { bail }` becomes
    `vmid_str.parse::<u32>().map_err(|_| anyhow!(...))?` — same semantics.
  * `if let Some(guest) = …find(…) { found = true; … } if !found { bail }`
    becomes `…find(…).ok_or_else(|| anyhow!(...))?` — same error path.
  * `if let Some(channel) = notify { if channel == "telegram" { ... } }`
    becomes `if notify.as_deref() == Some("telegram") { ... }`.

Public CLI surface is unchanged: `proxxx watch --help` still lists --since,
--target, --until, --timeout, --notify with identical semantics, and the
emitted JSON shapes (`condition_met` and the `diff` array) are byte-for-byte
the same. Verified by clippy + the existing test suite.
